ai_project_generator/requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12

# HTTP & Async Clients
httpx>=0.27.0
aiofiles>=24.1.0

# Utilities
python-dotenv>=1.0.1
click>=8.1.7
rich>=13.7.0
typer>=0.12.0
pyyaml>=6.0.1
toml>=0.10.2

# Logging
structlog>=24.2.0

# Authentication & Security
PyJWT>=2.9.0
cryptography>=43.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# System Monitoring & Performance
psutil>=6.1.0
memory-profiler>=0.61.0

# Data Validation & Serialization
marshmallow>=3.21.0
orjson>=3.10.0

# File Operations
pathlib2>=2.3.7; python_version < "3.4"
watchfiles>=0.21.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.7
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0
faker>=24.0.0

# Code Quality
black>=24.4.0
ruff>=0.4.0
mypy>=1.10.0
pylint>=3.2.0

# Type Hints
typing-extensions>=4.12.0

# Development Tools
ipython>=8.22.0
ipdb>=0.13.13

# Microservices & Serverless
redis>=5.0.0
celery>=5.3.0
rq>=1.15.0
arq>=0.25.0
pika>=1.3.0
kafka-python>=2.0.2

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-otlp>=1.21.0

# Serverless
mangum>=0.17.0

# Message Brokers (optional)
# rabbitmq-server required for RabbitMQ
# kafka required for Kafka

# Performance Optimizations
uvloop>=0.19.0; sys_platform != "win32"

# Deep Learning & AI Frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.25.0
peft>=0.7.0  # For LoRA and efficient fine-tuning
bitsandbytes>=0.41.0  # For quantization

# Gradio for Interactive Demos
gradio>=4.7.0

# Experiment Tracking
tensorboard>=2.15.0
wandb>=0.16.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
pillow>=10.1.0
opencv-python>=4.8.0

# Progress Bars
tqdm>=4.66.0

# Additional ML Utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional Dependencies for Advanced Features
huggingface-hub>=0.19.0  # For HF Hub integration
mlflow>=2.8.0  # For MLflow integration
onnxruntime>=1.16.0  # For ONNX export/import
nltk>=3.8.0  # For text preprocessing (stopwords)
matplotlib>=3.8.0  # For visualization
